move test data to test-data
