+++ packages/repocop/src/test-data/example-dependabot-alerts.ts
